requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.1.7
django-cors-headers==3.4.0
django-filter==21.1
django-rest-framework==0.1.0
django-simple-history==2.11.0
djangorestframework==3.11.0
djangorestframework-gis==0.18
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
pillow==10.3.0
psycopg2==2.9.3
psycopg2-binary==2.8.5
PyJWT==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2022.1
setproctitle==1.2.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
xlrd==2.0.1
